# public/openapi.yaml
openapi: 3.0.0
servers:
  - description: Localhost
    url: http://localhost:3000
info:
  version: "1.0.0"
  title: Motorola Face Detection API
  description: This api uses a ML algorithm to detect faces in an image
paths:
  /create:
    post:
      summary: Creates a new face detection request
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
        '400':
          description: Invalid Request
      requestBody:
        description: Image location for the request
        content:
          application/json:
            schema:
              type: object
              required:
                - image_location
              properties:
                image_location:
                  type: string
                  example: 'https://s.hdnux.com/photos/51/23/24/10827008/4/1200x0.jpg'
  /list:
    get:
      summary: Lists all face detection requests
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FaceDetection'
        '400':
          description: Invalid Request
  /status:
    get:
      summary: Get the status of a face detection request
      parameters:
        - name: id
          in: query
          required: true
          description: id of the detection request
          schema:
            type: integer
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceDetection'
        '404':
          description: Not Found
  /delete:
    delete: 
      summary: Deletes a face detection request
      parameters:
        - name: id
          in: query
          required: true
          description: id of the detection request
          schema:
            type: integer
      responses:
        '200':
          description: Successful deletion
        '400':
          description: Invalid Request
        '404':
          description: Not found
          
components:
  schemas:
    FaceDetection:
      type: object
      required:
        - id
        - status
        - image_location
      properties:
        id:
          type: integer
          example: 1
        status:
          type: string
          enum: ['pending', 'complete', 'failed']
          example: 'pending'
        face_count:
          type: integer
          example: null
        image_location:
          type: string
          example: 'http://path-to-image.jpeg'